+++ app/[lang]/(app)/(home)/(tablet)/restaurant/new-order/components/OrderItem.tsx
@@ -32,6 +32,7 @@ export default function OrderItem({
    order: { isLoading: userOrderIsLoading },
    orderItems: { isLoading: userOrderItemsLoading },
   },
+  accessibility,
  } = useOrderBaseConfigContext();
  const { format } = useCurrencyFormatter();
  const targetOrderItem = orderItems.filter(
@@ -96,7 +97,7 @@ export default function OrderItem({
         variant='ghost'
         size='icon-lg'
         className='text-primary rounded-full'
-        disabled={userOrderItemsLoading || userOrderIsLoading}
+        disabled={userOrderItemsLoading || userOrderIsLoading || !accessibility}
         onClick={() => {
          orderItemsDispatch({
           type: 'addOrderItems',
@@ -118,6 +119,7 @@ export default function OrderItem({
         variant='ghost'
         size='icon-lg'
         className='text-rose-600 dark:text-rose-400 rounded-full'
+        disabled={!accessibility}
         onClick={() => {
          orderItemsDispatch({
           type: 'decreaseOrderItemsAmount',
@@ -137,6 +139,7 @@ export default function OrderItem({
         variant='ghost'
         size='icon-lg'
         className='text-secondary rounded-full'
+        disabled={!accessibility}
         onClick={() => {
          orderItemsDispatch({
           type: 'increaseOrderItemsAmount',