+++ app/[lang]/(app)/(home)/(tablet)/restaurant/new-order/components/confirm-order/ConfrimOrderModal.tsx
@@ -40,6 +40,7 @@ export default function ConfirmOrderModal({
   closeConfirmOrder,
   shopInfoLoading,
   order: { orderItems, orderItemsDispatch, onCloseOrder, onSaveOrder },
+  accessibility,
  } = useOrderBaseConfigContext();
  return (
   <Dialog
@@ -149,7 +150,7 @@ export default function ConfirmOrderModal({
       <div className='flex gap-2 sm:gap-4'>
        {confirmOrderActiveType === 'shoppingCard' && !!orderItems.length && (
         <Button
-         disabled={shopInfoLoading}
+         disabled={shopInfoLoading || !accessibility}
          variant='destructive'
          className='h-11 max-sm:p-3 max-sm:grow'
          onClick={onCloseOrder}
@@ -162,7 +163,7 @@ export default function ConfirmOrderModal({
         <Button
          variant='secondary'
          className='h-11 max-sm:p-3 max-sm:grow'
-         disabled={shopInfoLoading}
+         disabled={shopInfoLoading || !accessibility}
          onClick={() => {
           onSaveOrder();
          }}