+++ app/[lang]/(app)/(home)/(tablet)/restaurant/new-order/components/order-invoice/OrderInvoice.tsx
@@ -71,6 +71,7 @@ export default function OrderInvoice({ dic }: { dic: NewOrderDictionary }) {
   shopLoading,
   confirmOrderActiveType,
   order: { orderItems, onCloseOrder, onSaveOrder },
+  accessibility,
   invoice: {
    isPayable,
    orderTotals: {
@@ -199,7 +200,9 @@ export default function OrderInvoice({ dic }: { dic: NewOrderDictionary }) {
    if (!walletInfoSuccess || !walletInfoData) {
     return (
      <Button
-      disabled={isFetching || shopLoading || getWalletInfoIsPending}
+      disabled={
+       isFetching || shopLoading || getWalletInfoIsPending || !accessibility
+      }
       type='submit'
       className='h-11'
       onClick={(e) => {
@@ -214,7 +217,7 @@ export default function OrderInvoice({ dic }: { dic: NewOrderDictionary }) {
    }
    return (
     <Button
-     disabled={isFetching || shopLoading}
+     disabled={isFetching || shopLoading || !accessibility}
      type='submit'
      className='h-11'
      onClick={(e) => {
@@ -237,7 +240,7 @@ export default function OrderInvoice({ dic }: { dic: NewOrderDictionary }) {
   if (paymentTypeValue?.key === '2') {
    return (
     <Button
-     disabled={isFetching || shopLoading}
+     disabled={isFetching || shopLoading || !accessibility}
      type='submit'
      className='h-11'
      onClick={handleConfirmPayment}
@@ -250,7 +253,7 @@ export default function OrderInvoice({ dic }: { dic: NewOrderDictionary }) {
   return (
    <Button
     type='submit'
-    disabled={isFetching || shopLoading}
+    disabled={isFetching || shopLoading || !accessibility}
     className='h-11'
     onClick={handleConfirmPayment}
    >
@@ -335,7 +338,7 @@ export default function OrderInvoice({ dic }: { dic: NewOrderDictionary }) {
     </div>
     <div className='grid grid-cols-2 sm:grid-cols-3 gap-3 mb-6 border-b border-input pb-6'>
      <Button
-      disabled={shopLoading}
+      disabled={shopLoading || !accessibility}
       variant='destructive'
       size='lg'
       className='font-medium disabled:bg-neutral-400 disabled:dark:bg-neutral-600'
@@ -345,7 +348,7 @@ export default function OrderInvoice({ dic }: { dic: NewOrderDictionary }) {
       {dic.invoice.closeOrder}
      </Button>
      <Button
-      disabled={shopLoading}
+      disabled={shopLoading || !accessibility}
       size='lg'
       className='font-medium'
       onClick={onSaveOrder}