+++ app/[lang]/(app)/(home)/(tablet)/restaurant/new-order/services/order-tools/OrderBaseConfigProvider.tsx
@@ -67,6 +67,7 @@ import {
  NewOrderSettings,
  saveNewOrderSettings,
 } from '../../utils/newOrderSettings';
+import { useUserAccessibilityContext } from '@/app/[lang]/(app)/login/services/user-accessibility/userAccessibilityContext';
 
 export default function OrderBaseConfigProvider({
  children,
@@ -75,6 +76,7 @@ export default function OrderBaseConfigProvider({
  children: ReactNode;
  dic: NewOrderDictionary;
 }) {
+ const { userAccessibility } = useUserAccessibilityContext();
  const queryClient = useQueryClient();
  const router = useRouter();
  const { locale } = useBaseConfig();
@@ -543,6 +545,7 @@ export default function OrderBaseConfigProvider({
   newOrderData: SaveOrderPackage['order'];
   orderInfoData: OrderInfo;
  } | null> {
+  if (!userAccessibility['restaurant']['newOrder']) return null;
   let newOrderData: SaveOrderPackage['order'] | null = null;
   let orderInfoData: OrderInfo | null = null;
   if (!initData) return null;
@@ -945,6 +948,7 @@ export default function OrderBaseConfigProvider({
    isLoading: systemPricingIsLoading,
    handleSetSystemPricing: onSetSystemPricing,
   },
+  accessibility: userAccessibility['restaurant']['newOrder'],
  };
 
  useEffect(() => {