+++ app/[lang]/(app)/login/hooks/useLogout.ts
@@ -1,3 +1,4 @@
+import { useCallback } from 'react';
 import { useRouter } from 'next/navigation';
 import { useBaseConfig } from '@/services/base-config/baseConfigContext';
 import { clearUserLoginToken } from '../utils/loginTokenManager';
@@ -9,10 +10,12 @@ export function useLogout() {
  const router = useRouter();
  const { locale } = useBaseConfig();
 
- return () => {
+ const logout = useCallback(() => {
   router.push(`/${locale}/login`);
   clearUserLoginToken();
   clearUserInfoRouterStorageValue();
   queryClient.clear();
- };
+ }, [locale, router, queryClient]);
+
+ return logout;
 }