+++ app/[lang]/(app)/login/services/userinfo-provider/utils/departmentRoutes.ts
@@ -1,4 +1,16 @@
+import { Departments } from './systems';
 const roomDevisionRoute = 'room-devision';
 const restaurantRotue = 'restaurant';
 
-export { roomDevisionRoute, restaurantRotue };
+function isTheRightPath(pathname: string, departmentID: number) {
+ const activeRoute = pathname.split('/').at(2);
+ switch (departmentID) {
+  case Departments.foodAndBeverage:
+   return activeRoute === restaurantRotue;
+  case Departments.roomDivision:
+   return activeRoute === roomDevisionRoute;
+ }
+ return true;
+}
+
+export { roomDevisionRoute, restaurantRotue, isTheRightPath };